dev_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking dev_requirements.in
#
alembic==1.9.1
    # via flask-migrate
astroid==2.12.13
    # via pylint
black==22.12.0
    # via -r dev_requirements.in
click==8.1.3
    # via
    #   black
    #   flask
dill==0.3.6
    # via pylint
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-sqlalchemy
flask-migrate==4.0.0
    # via -r requirements.in
flask-sqlalchemy==3.0.2
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==2.0.1
    # via sqlalchemy
isort==5.11.4
    # via
    #   -r dev_requirements.in
    #   pylint
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
lazy-object-proxy==1.8.0
    # via astroid
mako==1.2.4
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
mccabe==0.7.0
    # via pylint
mypy-extensions==0.4.3
    # via black
pathspec==0.10.3
    # via black
platformdirs==2.6.0
    # via
    #   black
    #   pylint
psycopg2==2.9.5
    # via -r requirements.in
pylint==2.15.9
    # via -r dev_requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
sqlalchemy==1.4.45
    # via
    #   alembic
    #   flask-sqlalchemy
tomlkit==0.11.6
    # via pylint
werkzeug==2.2.2
    # via flask
wrapt==1.14.1
    # via astroid
